requirements.txt
arrow==1.2.2
asgiref==3.5.2
attrs==21.4.0
beautifulsoup4==4.11.1
boto3==1.24.55
botocore==1.27.55
bravado==11.0.3
cachetools==5.2.0
Cerberus==1.3.4
certifi==2022.6.15
charset-normalizer==2.0.12
decouple==0.0.7
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.13.1
ERPpeek==1.7.1
filelock==3.7.1
fqdn==1.5.1
gunicorn==20.1.0
idna==3.3
isoduration==20.11.0
jmespath==1.0.1
jsonpointer==2.3
jsonref==0.2
jsonschema==4.6.0
lxml==4.9.1
monotonic==1.6
MouseInfo==0.1.3
msgpack==1.0.4
OdooRPC==0.8.0
Pillow==9.1.1
pipenv==2022.6.7
platformdirs==2.5.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
python-memcached==1.59
pytweening==1.0.4
pytz==2022.1
pywin32==304
PyYAML==6.0
requests==2.28.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
swagger-spec-validator==2.4.3
typing_extensions==4.2.0
tzdata==2022.1
uri-template==1.2.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
webcolors==1.12
whitenoise==6.2.0
